Extract graph query runner partial
